refactor: decompose ArtworkGallery.tsx into focused subcomponents
Extracts LightboxDialog, GalleryGrid, ProjectSwitcher, CollectionSection,
and BrewInstallBlock. No behavioral changes.
